[Refactor] 타워 데미지 관련 int -> flaot으로 변경

+++ Assets/02.Script/Data/Tower/Attack/BeamAttackData.cs
@@ -6,7 +6,7 @@ public class BeamAttackData : AttackData
 {
     public float stackInterval;
     public int maxStack;
-    public int beamDamagePerStack;
+    public float beamDamagePerStack;
     public GameObject beamPrefab;
 
     public override ITowerAttack CreateAttack(Tower tower, List<ITowerEffect> effects)
