+++ Assets/02.Script/Data/Tower/Effect/DamageEffectData.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "Tower/Effect/Damage")]
 public class DamageEffectData : EffectData
 {
-    public int damage;
-    public int damageGrowth; // 강화 1회당 증가량
+    public float damage;
+    public float damageGrowth; // 강화 1회당 증가량
 
     public override ITowerEffect CreateEffect(Tower tower)
     {
